Transforms the tips and targets so things don't flip-flop
Subtract the origin position in the update_tip_headings and update_target_headings code.

The coordinate space of the goal can depend on what the user intends. They need to be converted to the skeleton parent coordinate space or global space before the solver works with them.

We might want to define the goal relative to the skeleton so that a character tumbling through space can still move their arms to target parts on their own body. 

While in another use case, we might want to set a goal relative to the world so that a character can keep their foot on the exact ground location as they walk.