Add support for physical bone in ManyBoneIK3D
This commit adds support for a new type of node called PhysicalBone3D in the ManyBoneIK3D module. The create_pin_target_node function has been modified to create and use PhysicalBone3D instead of Node3D when finding and setting the pin target node. This change allows for more accurate physics simulation and interaction with the skeleton.